feature(Libpfm): Libpfm sensors implementation
Implements the Libpfm's class to access to the performance counters (+ jar jnaerated),
and the Libpfm's sensors.

+++ src/main/scala/org/powerapi/core/OSHelper.scala
@@ -31,7 +31,7 @@ import org.powerapi.configuration.LogicalCoresConfiguration
  *
  * @author Maxime Colmant <maxime.colmant@gmail.com>
  */
-case class Thread(tid: Long)
+case class Thread(tid: Int)
 
 /**
  * Wrapper class for the time spent by the cpu in each frequency (if dvfs enabled).
@@ -157,7 +157,7 @@ class LinuxHelper extends OSHelper with Configuration with LogicalCoresConfigura
 
   def getProcesses(application: Application): List[Process] = {
     Seq("ps", "-C", application.name, "-o", "pid", "--no-headers").!!.split("\n").toList.map {
-      case PSFormat(pid) => Process(pid.toLong)
+      case PSFormat(pid) => Process(pid.toInt)
     }
   }
 
@@ -168,7 +168,7 @@ class LinuxHelper extends OSHelper with Configuration with LogicalCoresConfigura
       /**
        * The pid is removed because it corresponds to the main thread.
        */
-      pidDirectory.listFiles.filter(dir => dir.isDirectory && dir.getName != s"${process.pid}").toList.map(dir => Thread(dir.getName.toLong))
+      pidDirectory.listFiles.filter(dir => dir.isDirectory && dir.getName != s"${process.pid}").toList.map(dir => Thread(dir.getName.toInt))
     }
     else List()
   }
